refactor(ratelimit,budget): serve both SQL backends from one store
Both stores carry a config-vs-manual precedence upsert and a config-replace
transaction; those are now written once. Their pre-scope and add-column
migrations unify too — only the column introspection differs per engine
(PRAGMA versus information_schema), so the rebuild statements are shared and
just the lookup switches.

Each file also held the same prepared-statement upsert twice, once inline and
once in a helper; that collapses to one.

budget's SumUsageCost stays dialect-specific and is marked as such: SQLite
keeps the usage timestamp as text and must convert it, while PostgreSQL
compares a real timestamp column and has to quote "usage" as a reserved word.

The rate-limit pre-scope migration now runs on PostgreSQL in tests for the
first time.
